requirements.txt
autopep8==1.4.3
Babel==2.6.0
Django==1.11.18
django-phonenumber-field==2.2.0
migrations==0.1.1
phonenumbers==8.10.7
Pillow==5.4.1
pycodestyle==2.5.0
pytz==2018.9
